Added uploading image functionality

+++ routes/stories.js
@@ -2,6 +2,7 @@ const express = require("express");
 const router = express.Router();
 const { ensureAuth } = require("../middleware/auth");
 const Story = require("../models/Story");
+
 const {
   truncate,
   stripTags,
@@ -10,6 +11,23 @@ const {
   formatDate,
   escapeRegExp,
 } = require("../helpers/ejs");
+
+const multer=require("multer");
+
+const {storage, cloudinary}=require("../config/cloudinary");
+
+const upload=multer({storage,
+  fileFilter: function (req, file, cb) {
+  if (!file.originalname.match(/\.(jpg|jpeg|png)$/)) {
+    return cb(new Error('Only images of type jpg, jpeg, or png are allowed'));
+  }
+  cb(null, true);
+},
+limits: { fileSize: 2000000 }
+})
+
+const {uploadFile,editFile}=require("../helpers/uploadFile");
+
 
 // @desc Show add page
 // @route GET /stories/add
@@ -80,19 +98,22 @@ router.get("/:id", ensureAuth, async (req, res) => {
   }
 });
 
-// @desc Process the add form
+//@desc Process the add form
 // @route Post /stories
-router.post("/", ensureAuth, async (req, res) => {
+router.post("/", ensureAuth, uploadFile, async (req, res) => {
   try {
     req.body.user = req.user.id;
-    console.log(req.body);
-    await Story.create(req.body);
+    req.body.images=req.files.map(f=>({url:f.path, fileName:f.filename}));
+    const story=new Story(req.body);
+    await story.save();
     res.redirect("/dashboard");
   } catch (err) {
     console.log(err);
     res.render("error", { err, layout: "layouts/boilerplate" });
   }
 });
+
+
 
 // @desc Show all stories
 // @route GET /stories
@@ -130,6 +151,8 @@ router.get("/edit/:id", ensureAuth, async (req, res) => {
     if (req.user.id != story.user) {
       return res.redirect("/stories");
     } else {
+
+
       return res.render("stories/edit", {
         story,
         statusList,
@@ -143,8 +166,8 @@ router.get("/edit/:id", ensureAuth, async (req, res) => {
 });
 
 // @desc Process the edited form
-// @route PUT /stories/edit/:id
-router.put("/edit/:id", ensureAuth, async (req, res) => {
+// @route PUT /stories/:id
+router.put("/:id", ensureAuth, editFile,async (req, res) => {
   try {
     let story = await Story.findById(req.params.id);
 
@@ -154,11 +177,23 @@ router.put("/edit/:id", ensureAuth, async (req, res) => {
     if (req.user.id != story.user) {
       return res.redirect("/stories");
     } else {
+      if(req.body.deleteImages)
+      {
+        for(let fileName of req.body.deleteImages){
+          await cloudinary.uploader.destroy(fileName);
+        }
+        await story.updateOne({$pull:{images:{fileName:{$in:req.body.deleteImages}}}});
+    
+      }
+
+    
       story = await Story.findByIdAndUpdate({ _id: req.params.id }, req.body, {
-        new: true,
+
         runValidators: true,
       });
-
+      const imgs=req.files.map(f=>({url:f.path, fileName:f.filename}));
+      story.images.push(...imgs);
+      await story.save();
       res.redirect("/dashboard");
     }
   } catch (err) {
@@ -171,6 +206,16 @@ router.put("/edit/:id", ensureAuth, async (req, res) => {
 //@route DELETE /stories/id
 router.delete("/:id", ensureAuth, async (req, res) => {
   try {
+    let story = await Story.findById(req.params.id);
+    if (!story) {
+      return res.render("404error", { layout: "layouts/boilerplate" });
+    }
+
+    console.log(story);
+    for(let file of story.images){
+      await cloudinary.uploader.destroy(file.fileName);
+    }
+
     await Story.deleteOne({ _id: req.params.id });
     res.redirect("/dashboard");
   } catch (err) {
